Natasha/Pat #3 Update header and footer in main and translation activities. Change Activity names.
