+++ app/src/main/AndroidManifest.xml
@@ -11,13 +11,13 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
-        <activity android:name=".CollectionsActivity" >
+        <activity android:name=".MainActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"/>
+        <activity android:name=".TranslationsActivity"/>
         <activity android:name=".RecordingActivity" />
         <activity android:name=".ImportActivity">
             <!-- To support opening from a file manager. -->
