+++ app/src/main/res/layout/activity_main.xml
@@ -1,53 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:background="#f2f2f2"
-    android:id="@+id/mainActivity">
-
-    <HorizontalScrollView
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:fillViewport="true"
-        android:id="@+id/tabScroll">
-        <LinearLayout
-            android:orientation="horizontal"
-            android:layout_width="wrap_content"
-            android:layout_height="56dp"
-            android:background="@color/colorPrimary"
-            android:id="@+id/tabs" />
-    </HorizontalScrollView>
+    android:background="@color/listBackgroundColor">
 
     <ListView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:divider="#f2f2f2"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:divider="@color/listBackgroundColor"
         android:dividerHeight="20dp"
-        android:layout_below="@+id/tabScroll"
-        android:id="@+id/list" />
-
-    <ImageView
-        android:layout_width="fill_parent"
-        android:layout_height="8dp"
-        android:background="#00000000"
-        android:src="@drawable/main_tabbar_shadow"
-        android:scaleType="fitXY"
-        android:layout_below="@+id/tabScroll"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:id="@+id/tabs_shadow" />
-
-    <ImageButton
-        android:layout_width="74dp"
-        android:layout_height="78dp"
-        android:background="@drawable/add_button"
-        android:id="@+id/add_button"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginRight="12dp"
-        android:layout_marginBottom="4dp" />
+        android:id="@+id/collections_list"/>
 </RelativeLayout>
