+++ app/src/main/res/layout/collections_footer.xml
@@ -1,10 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:paddingBottom="80dp"
     android:orientation="vertical">
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="center_horizontal"
+        android:layout_marginTop="6dp"
+        android:text="@string/main_feedback_request" />
     <Button
         android:layout_width="wrap_content"
         android:layout_height="36dp"
@@ -15,6 +20,6 @@
         android:textSize="14dp"
         android:textColor="@color/textColor"
         android:background="@color/colorPrimary"
-        android:text="@string/main_export_button_text"
-        android:onClick="onExportButtonPress" />
+        android:text="@string/main_feedback_button_text"
+        android:id="@+id/feedback_button" />
 </LinearLayout>
